% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section10.R
\name{all_ancova}
\alias{all_ancova}
\title{Generate ANCOVA Analysis}
\usage{
all_ancova(...)
}
\arguments{
\item{input}{A data frame, or three vectors corresponding to (x, y, treatment)}
}
\value{
Returns the ANCOVA Table for the inputs.
}
\description{
This function performs an ANCOVA analysis based on the input data. We construct a scalar ANCOVA model of the following form:
\deqn{y_{ij} = \eta + \tau_i + \beta (x_{ij} - \bar{x}_{..}) + \epsilon_{ij}}
}
\details{
Where:
\itemize{
\item \eqn{y_{i,j}}{y_{ij}} is the \eqn{i,j}{i,j}-th response.
\item \eqn{\eta}{\eta} is the Grand Mean.
\item \eqn{\tau_i}{\tau_i} is the effect of being assigned treatment \eqn{i}{i}.
\item \eqn{\beta}{\beta} is the regression coefficient for the covariate.
\item \eqn{x_{i, j}}{x_{ij}} is the covariate associated with \eqn{y_{i,j}}{y_{ij}}.
\item \eqn{\bar{x}_{..}}{\bar{x}_{..}} is the overall mean of \eqn{x}{x} observations.
\item \eqn{\epsilon_{i,j}}{\epsilon_{ij}} is assumed to be iid \eqn{N(0, \sigma^2)}{N(0, \sigma^2)} with \eqn{\sigma^2}{\sigma^2} the same for all levels.
}
}
\examples{
## EXAMPLE 1 : Three Inputs: x, y, and treatment
all_ancova(x = iris$Sepal.Length,
           y = iris$Sepal.Width,
           treat = iris$Species)
## EXAMPLE 2 : Data Frame,  x, y, treatment inferred
all_ancova( iris[3:5] )
}
