% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section8.R
\name{explode_word}
\alias{explode_word}
\title{Explode Words into Characters}
\usage{
explode_word(words)
}
\arguments{
\item{words}{A character vector, where each element is a word that you want to split into characters.}
}
\value{
A list where each element is a character vector corresponding to the characters of a word.
}
\description{
This function takes a vector of words and splits each word into its constituent characters.
}
\examples{
explode_word(c("tomato"))
explode_word(c("apple", "banana"))
}
