% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section7.R
\name{generate_normal_plot}
\alias{generate_normal_plot}
\title{Generate Normal or Half-Normal Plot of Estimated Effects}
\usage{
generate_normal_plot(y, half_normal = FALSE, plus = FALSE)
}
\arguments{
\item{y}{Numeric vector representing the input data for which the estimated effects are to be plotted.
The length of \code{y} must be a power of 2, aligning with the factorial design in ANOVA.}

\item{half_normal}{Logical; if TRUE, creates a half-normal plot, otherwise, a normal plot is generated.}

\item{plus}{Logical; if TRUE, uses positive-first ordering in generating the planning matrix.
This parameter is passed to \code{\link{generate_planning_matrix}}.}
}
\value{
A ggplot object representing the normal or half-normal plot of estimated effects.
}
\description{
This function generates a normal or half-normal plot of estimated effects based on input data.
It's useful for analyzing the effects in a multi-factor ANOVA design. The function computes
estimated effects using a planning matrix, which is generated by the
\code{\link{generate_planning_matrix}} function, and compares them against the theoretical
quantiles of the standard normal distribution.
}
\examples{
# Create dummy data
set.seed(404)
dummy_y <- rnorm(2^4)

# Generate a normal plot
generate_normal_plot(dummy_y)

# Generate a half-normal plot
generate_normal_plot(dummy_y, half_normal = TRUE)
}
\seealso{
\code{\link{generate_planning_matrix}} for the function used to generate the planning matrix.
}
