% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section7.R
\name{generate_planning_matrix}
\alias{generate_planning_matrix}
\title{Generate Planning Matrix for Multi-factor ANOVA}
\usage{
generate_planning_matrix(k, plus_first = TRUE)
}
\arguments{
\item{k}{An integer indicating the number of factors in the ANOVA design.}

\item{plus_first}{Logical; if TRUE, positive entries precede negative ones in the planning matrix.
If FALSE, the order is reversed. Normally, in STAT 404, this should be FALSE.
However, in other works it is normally TRUE.}
}
\value{
A data frame representing the planning matrix, with columns for each factor and their interactions.
The columns are named using uppercase letters starting from 'A'.
}
\description{
This function generates a planning matrix used in the context of multi-factor ANOVA experimental design.
It creates a matrix based on the number of factors (k) and the order of entries. The matrix includes
all combinations of factor levels and their interactions.
}
\examples{
# Generate a planning matrix for 3 factors with positive entries first
generate_planning_matrix(3, plus_first = TRUE)

# Generate a planning matrix for 4 factors with negative entries first
generate_planning_matrix(4, plus_first = FALSE)
}
