% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section8.R
\name{multiply_words}
\alias{multiply_words}
\title{Multiply Two Words}
\usage{
multiply_words(word1, word2, exploded = FALSE)
}
\arguments{
\item{word1}{The first word (a character string or a vector of characters if \code{exploded} is TRUE).}

\item{word2}{The second word (a character string or a vector of characters if \code{exploded} is TRUE).}

\item{exploded}{Logical, indicating whether the words are already split into characters.}
}
\value{
A sorted character vector containing the unique letters from both words, minus duplicates.
}
\description{
This function takes the "product" of two words, combining their letters and dropping duplicates.
It optionally uses the previously defined \code{\link{explode_word}} function to split the words into characters.
}
\examples{
multiply_words(c("A", "B"), c("A", "D"), TRUE) # returns c("B", "D")
multiply_words("Tuhmaytoe", "Tuhmaatoe") # returns c("y")
}
\seealso{
\code{\link{explode_word}} for the function used to split words into characters.
}
