% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section8.R
\name{all_contrast_subgroups}
\alias{all_contrast_subgroups}
\title{Generate All Contrast Subgroups}
\usage{
all_contrast_subgroups(def_relations)
}
\arguments{
\item{def_relations}{A character vector representing different definitions or
relationships from which the contrast subgroups are to be generated.}
}
\value{
A list of character vectors, each representing a unique contrast subgroup,
ordered by the length of the subgroup. The list always starts with "I" to denote
the identity or baseline subgroup. Works well with \code{\link{explode_word}} to translate
a list of characters into a suitable list of lists.
}
\description{
This function generates all possible unique contrast subgroups (combinations) from
a given set of definitions or relationships. It's designed to explore all potential
arrangements and their unique compositions, particularly useful in statistical
analysis or fractional factorial designs.
}
\examples{
IA = c("1235", "1246", "12347")
all_contrast_subgroups(explode_word(IA))
}
\seealso{
\code{\link{explode_word}} for the function used to generate the planning matrix.
}
